refactor(tenant-api): split rate limiter out of middleware.go into rate_limiter.go
middleware.go (490 lines) mixed four unrelated concerns: the per-caller
sliding-window rate limiter (~330 lines), request-ID echo, structured request
logging, and request-body size config. Move all rate-limiter code (RateLimit,
rateLimiter + methods, sweeper, RateLimitConfig[FromEnv], rateLimitCaller,
activeLimiter, RateLimitMetrics) into a new same-package file rate_limiter.go.
middleware.go now holds only RequestIDResponse, SlogRequestLogger, and the
body-size config (490 -> 128 lines).

Pure intra-package file move: same package handler, no caller/test changes.
Adversarial review confirms the 354-line moved region is character-identical,
the symbol set is conserved (declared once each), imports are exact, and
go build / vet / test all pass.

Self-Review-Pass-2: adversarial review confirms faithful move; symbols conserved; suite green
